pyproject.toml
[project]
name = "rcm-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "appwrite>=14.1.0",
    "fastapi[standard]>=0.128.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.45",
    "timing-asgi>=0.3.1",
    "ulid>=1.1",
    "uvicorn>=0.40.0",
]
